requirements.txt
anthropic
python-dotenv
aiohttp